MAILET-115 Move message altering from AbstractRedirect to AlteredMailUtils
