+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Forward.java
@@ -106,7 +106,7 @@ public class Forward extends AbstractRedirect {
     }
 
     @Override
-    protected String getMessage(Mail originalMail) throws MessagingException {
+    public String getMessage(Mail originalMail) throws MessagingException {
         return getInitParameters().getMessage();
     }
