+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/NotifySender.java
@@ -143,7 +143,7 @@ public class NotifySender extends AbstractRedirect {
     }
 
     @Override
-    protected String getMessage(Mail originalMail) throws MessagingException {
+    public String getMessage(Mail originalMail) throws MessagingException {
         return new NotifyMailetsMessage().generateMessage(getInitParameters().getMessage(), originalMail);
     }
