+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Redirect.java
@@ -315,7 +315,7 @@ public class Redirect extends AbstractRedirect {
     }
 
     @Override
-    protected String getMessage(Mail originalMail) throws MessagingException {
+    public String getMessage(Mail originalMail) throws MessagingException {
         return getInitParameters().getMessage();
     }
