+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Resend.java
@@ -308,7 +308,7 @@ public class Resend extends AbstractRedirect {
     }
 
     @Override
-    protected String getMessage(Mail originalMail) throws MessagingException {
+    public String getMessage(Mail originalMail) throws MessagingException {
         return getInitParameters().getMessage();
     }
